+++ code/test/sixsq/nuvla/server/util/response_test.clj
@@ -2,14 +2,14 @@
   (:require
     [clojure.string :as str]
     [clojure.test :refer [are deftest is]]
-    [sixsq.nuvla.server.util.response :as r])
+    [sixsq.nuvla.server.util.response :as t])
   (:import
     (clojure.lang ExceptionInfo)))
 
 
 (deftest check-response-created
   (let [id "RESOURCE_ID"
-        r  (r/response-created id)]
+        r  (t/response-created id)]
     (is (= 201 (:status r)))
     (is (= id (get-in r [:headers "Location"])))
     (is (nil? (:cookies r))))
@@ -17,7 +17,7 @@
   (let [id           "RESOURCE_ID"
         cookie-name  "MY_COOKIE"
         cookie-value "MY_COOKIE_VALUE"
-        r            (r/response-created id [cookie-name cookie-value])]
+        r            (t/response-created id [cookie-name cookie-value])]
     (is (= 201 (:status r)))
     (is (= id (get-in r [:headers "Location"])))
     (is (= cookie-value (get-in r [:cookies cookie-name])))))
@@ -25,7 +25,7 @@
 
 (deftest check-response-final-redirect
   (let [location "collection/my-new-resource"
-        r        (r/response-final-redirect location)]
+        r        (t/response-final-redirect location)]
     (is (= 303 (:status r)))
     (is (= location (get-in r [:headers "Location"])))
     (is (nil? (:cookies r))))
@@ -33,7 +33,7 @@
   (let [location     "RESOURCE_ID"
         cookie-name  "MY_COOKIE"
         cookie-value "MY_COOKIE_VALUE"
-        r            (r/response-final-redirect location [cookie-name cookie-value])]
+        r            (t/response-final-redirect location [cookie-name cookie-value])]
     (is (= 303 (:status r)))
     (is (= location (get-in r [:headers "Location"])))
     (is (= cookie-value (get-in r [:cookies cookie-name])))))
@@ -41,7 +41,7 @@
 
 (deftest check-json-response
   (let [body     {:key "value"}
-        response (r/json-response body)]
+        response (t/json-response body)]
     (is (= 200 (:status response)))
     (is (= body (:body response)))
     (is (= "application/json" (get-in response [:headers "Content-Type"])))))
@@ -53,7 +53,7 @@
         id       "collection/resource-id"
         location "collection/new-resource-id"]
 
-    (are [args expected] (= expected (apply r/map-response args))
+    (are [args expected] (= expected (apply t/map-response args))
 
                          [msg status]
                          {:status  status
@@ -84,11 +84,11 @@
         id       "collection/resource-id"
         location "collection/new-resource-id"]
 
-    (let [ex (r/ex-response msg status)]
+    (let [ex (t/ex-response msg status)]
       (is (instance? ExceptionInfo ex))
       (is (= msg (.getMessage ex))))
 
-    (are [args expected] (= expected (ex-data (apply r/ex-response args)))
+    (are [args expected] (= expected (ex-data (apply t/ex-response args)))
 
                          [msg status]
                          {:status  status
@@ -115,7 +115,7 @@
 
 (deftest check-response-deleted
   (let [id       "collection/resource-id"
-        response (r/response-deleted id)
+        response (t/response-deleted id)
         msg      (str id " deleted")]
     (is (= 200 (:status response)))
     (is (= msg (-> response :body :message)))))
@@ -123,7 +123,7 @@
 
 (deftest check-response-updated
   (let [id       "collection/resource-id"
-        response (r/response-updated id)
+        response (t/response-updated id)
         msg      (str "updated " id)]
     (is (= 200 (:status response)))
     (is (= msg (-> response :body :message)))))
@@ -131,7 +131,7 @@
 
 (deftest check-response-not-found
   (let [id       "collection/resource-id"
-        response (r/response-not-found id)
+        response (t/response-not-found id)
         msg      (str id " not found")]
     (is (= 404 (:status response)))
     (is (= msg (-> response :body :message)))))
@@ -139,33 +139,33 @@
 
 (deftest check-response-error
   (let [msg      "BAD THING HAPPENED"
-        response (r/response-error msg)]
+        response (t/response-error msg)]
     (is (= 500 (:status response)))
     (is (pos? (str/index-of (-> response :body :message) msg)))))
 
 
 (deftest check-response-conflict
   (let [id       "collection/resource-id"
-        response (r/response-conflict id)
+        response (t/response-conflict id)
         msg      (str "conflict with " id)]
     (is (= 409 (:status response)))
     (is (= msg (-> response :body :message)))))
 
 
 (deftest check-ex-bad-request
-  (let [response (ex-data (r/ex-bad-request))]
+  (let [response (ex-data (t/ex-bad-request))]
     (is (= 400 (:status response)))
     (is (= "invalid request" (-> response :body :message)))
 
     (let [msg      "bad-thing"
-          response (ex-data (r/ex-bad-request msg))]
+          response (ex-data (t/ex-bad-request msg))]
       (is (= 400 (:status response)))
       (is (= msg (-> response :body :message))))))
 
 
 (deftest check-ex-not-found
   (let [id       "collection/resource-id"
-        response (ex-data (r/ex-not-found id))
+        response (ex-data (t/ex-not-found id))
         msg      (str id " not found")]
     (is (= 404 (:status response)))
     (is (= msg (-> response :body :message)))))
@@ -173,7 +173,7 @@
 
 (deftest check-ex-conflict
   (let [id       "collection/resource-id"
-        response (ex-data (r/ex-conflict id))
+        response (ex-data (t/ex-conflict id))
         msg      (str "conflict with " id)]
     (is (= 409 (:status response)))
     (is (= msg (-> response :body :message)))))
@@ -181,11 +181,11 @@
 
 (deftest check-ex-unauthorized
   (let [id       "collection/resource-id"
-        response (ex-data (r/ex-unauthorized id))]
+        response (ex-data (t/ex-unauthorized id))]
     (is (= 403 (:status response)))
     (is (pos? (str/index-of (-> response :body :message) id))))
 
-  (let [response (ex-data (r/ex-unauthorized nil))]
+  (let [response (ex-data (t/ex-unauthorized nil))]
     (is (= 403 (:status response)))
     (is (= "credentials required" (-> response :body :message)))))
 
@@ -193,20 +193,20 @@
 (deftest check-ex-bad-method
   (let [uri      "collection/resource-id"
         method   "HEAD"
-        response (ex-data (r/ex-bad-method {:uri uri, :request-method method}))]
+        response (ex-data (t/ex-bad-method {:uri uri, :request-method method}))]
 
     (is (= 405 (:status response)))
     (is (zero? (str/index-of (-> response :body :message) "invalid method (")))
     (is (pos? (str/index-of (-> response :body :message) uri)))
     (is (pos? (str/index-of (-> response :body :message) method)))
 
-    (let [response (ex-data (r/ex-bad-method {:uri uri}))]
+    (let [response (ex-data (t/ex-bad-method {:uri uri}))]
       (is (= 405 (:status response)))
       (is (zero? (str/index-of (-> response :body :message) "invalid method (")))
       (is (pos? (str/index-of (-> response :body :message) uri)))
       (is (nil? (str/index-of (-> response :body :message) method))))
 
-    (let [response (ex-data (r/ex-bad-method {:request-method method}))]
+    (let [response (ex-data (t/ex-bad-method {:request-method method}))]
       (is (= 405 (:status response)))
       (is (zero? (str/index-of (-> response :body :message) "invalid method (")))
       (is (nil? (str/index-of (-> response :body :message) uri)))
@@ -217,7 +217,7 @@
   (let [uri      "collection/resource-id"
         method   "HEAD"
         action   "DO_IT"
-        response (ex-data (r/ex-bad-action {:uri uri, :request-method method} action))]
+        response (ex-data (t/ex-bad-action {:uri uri, :request-method method} action))]
 
     (is (= 404 (:status response)))
     (is (zero? (str/index-of (-> response :body :message) "undefined action (")))
@@ -225,21 +225,21 @@
     (is (pos? (str/index-of (-> response :body :message) method)))
     (is (pos? (str/index-of (-> response :body :message) action)))
 
-    (let [response (ex-data (r/ex-bad-action {:uri uri} action))]
+    (let [response (ex-data (t/ex-bad-action {:uri uri} action))]
       (is (= 404 (:status response)))
       (is (zero? (str/index-of (-> response :body :message) "undefined action (")))
       (is (pos? (str/index-of (-> response :body :message) uri)))
       (is (nil? (str/index-of (-> response :body :message) method)))
       (is (pos? (str/index-of (-> response :body :message) action))))
 
-    (let [response (ex-data (r/ex-bad-action {:request-method method} action))]
+    (let [response (ex-data (t/ex-bad-action {:request-method method} action))]
       (is (= 404 (:status response)))
       (is (zero? (str/index-of (-> response :body :message) "undefined action (")))
       (is (nil? (str/index-of (-> response :body :message) uri)))
       (is (pos? (str/index-of (-> response :body :message) method)))
       (is (pos? (str/index-of (-> response :body :message) action))))
 
-    (let [response (ex-data (r/ex-bad-action {:uri uri, :request-method method} nil))]
+    (let [response (ex-data (t/ex-bad-action {:uri uri, :request-method method} nil))]
       (is (= 404 (:status response)))
       (is (zero? (str/index-of (-> response :body :message) "undefined action (")))
       (is (pos? (str/index-of (-> response :body :message) uri)))
@@ -249,7 +249,7 @@
 
 (deftest check-ex-bad-CIMI-filter
   (let [err      {:key "value"}
-        response (ex-data (r/ex-bad-CIMI-filter err))]
+        response (ex-data (t/ex-bad-CIMI-filter err))]
     (is (= 400 (:status response)))
     (is (pos? (str/index-of (-> response :body :message) (pr-str err))))))
 
@@ -258,9 +258,46 @@
   (let [msg          "MESSAGE"
         id           "collection/resource-id"
         redirect-uri "somewhere/else"
-        response     (ex-data (r/ex-redirect msg id redirect-uri))]
+        response     (ex-data (t/ex-redirect msg id redirect-uri))]
     (is (= 303 (:status response)))
     (is (zero? (str/index-of (-> response :body :message) msg)))
     (let [location (get-in response [:headers "Location"])]
       (is (nat-int? (str/index-of location redirect-uri)))
       (is (nat-int? (str/index-of location msg))))))
+
+(deftest rethrow-response
+  (let [id "resource/id"]
+    (are [expect response]
+      (= expect (try
+                  (t/rethrow-response response)
+                  (catch Exception e
+                    (ex-data e))))
+      {:body    {:message     "resource/id not found"
+                 :resource-id "resource/id"
+                 :status      404}
+       :headers {"Content-Type" "application/json"}
+       :status  404} (ex-data (t/ex-not-found id))
+      {:body    {:message "bad request error"
+                 :status  400}
+       :headers {"Content-Type" "application/json"}
+       :status  400} (ex-data (t/ex-bad-request "bad request error"))
+      {:body    {:message "unexpected error occurred: rethrow-response bad argument"
+                 :status  500}
+       :headers {"Content-Type" "application/json"}
+       :status  500} (t/response-created id))))
+
+(deftest status-200?
+  (are [expect response]
+    (= expect (t/status-200? response))
+    true {:status 200}
+    false {:status 100}
+    false {:status 201}
+    false {:status 500}
+    false {:status nil}))
+
+(deftest throw-response-not-200
+  (is (= {:any "value"
+          :status 200} (t/throw-response-not-200 {:any "value"
+                                                  :status 200})))
+  (is (thrown? Exception (t/throw-response-not-200 {:status 100})))
+  (is (thrown? Exception (t/throw-response-not-200 nil))))